sweet: rootdir: Check build date before remove settings cache and persistent_properties
* It will always remove settings cache and persistent_properties.
  So, compare preeviously saved ro.system.build.date.utc value with current value

+++ rootdir/bin/init.qcom.sh
@@ -51,5 +51,19 @@ fi
 chmod g-w /data/vendor/modem_config
 setprop ro.vendor.ril.mbn_copy_completed 1
 
-# Clear packages_cache to avoid systemapp have mismatch strings id
-rm -rf /data/system/package_cache
+#
+# Save build date property as file
+#
+if [ ! -f /data/system/build_date.txt ]; then
+    getprop ro.system.build.date.utc>/data/system/build_date.txt
+fi
+build_date_file=`cat /data/system/build_date.txt`
+build_date_prop=`getprop ro.system.build.date.utc`
+# Compare build date before clean-up
+if [ $build_date_prop != $build_date_file ]; then
+    # Remove settings cache and persistent_properties to avoid derps after dirty flash
+    rm -rf /data/system/package_cache
+    # Remove and update build date file
+    rm -rf /data/system/build_date.txt
+    getprop ro.system.build.date.utc>/data/system/build_date.txt
+fi
